// angular/src/app/pages/analytics/type.ts
export interface MilestoneEstimationHistoryData {
    left_milestones: number;
    complete_date: string;
    created_at: string;
    id: string;
    milestone_id: string;
    optimistic_estimated_complete_date: string;
    estimation_date: string;
    pessimistic_estimated_complete_date: string;
    left_tasks: number;
}